+++ unlock-js/src/typedData/keyMetadata.js
